+++ app/src/main/java/com/sopt/now/compose/domain/entity/UserEntity.kt
@@ -1,0 +1,8 @@
+package com.sopt.now.compose.domain.entity
+
+data class UserEntity(
+    val id: String,
+    val pw: String,
+    val nickname: String,
+    val tel: String,
+)
